force ncurses mode if supported when hex colors are set in config
instead of exiting with error it is better to force ncurses on.
noncurses is now default, but does not support hex colors. so if
hex colors are used (like in a gradient) it will be better to force
ncurses mode. ncurses might have some performance issues and slight
glitches. but it should be ok as long as it is documented in the
config file. relates to #368